requirements.txt
pytz==2022.2
vk-api==11.9.9
pandas==1.5
click==8.1.3
python-dotenv==0.21.0